聚合根基类、实体基类、值类型基类封装

+++ Repository/UserRepository.cs
@@ -6,7 +6,7 @@ using System.Text;
 
 namespace Repository
 {
-    public class UserRepository : Repository<User, long>, IUserRepository
+    public class UserRepository : Repository<User, int>, IUserRepository
     {
     }
 }
